Made websocket proxy more protocol specific
The main reason for the WebSocket proxy was to enable browser
clients to talk with backend servers using the 9P protocol.

This version makes that more explicity and remove the control messages
sent using the websocket connection.

The next change is to return Rerror messages to the client if something
happens with the underlying connection, and ensure that we clean all
pending fids if the client disconnects without a proper clean-up.